% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delempcol-package.R
\docType{package}
\name{delempcol-package}
\alias{delempcol-package}
\title{delempcol: Delete Empty Columns from a Dataset}
\description{
With the delempcol package, you can delete all columns that exist only of missing values from the dataset at once.
}
\keyword{internal}
