% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delempcol.R
\name{delempcol}
\alias{delempcol}
\title{Delete Empty Columns from a Dataset}
\usage{
delempcol(data)
}
\arguments{
\item{data}{A dataframe you want to delete all the empty columns from.}
}
\value{
A dataframe without empty columns.
}
\description{
With the delempcol package, you can delete all columns that exist only of
missing values from the dataset at once.
}
\examples{
library(dplyr)
data.frame(full_column = c(1,2,3,4,5), empty_column = c(NA, NA, NA, NA, NA)) \%>\% delempcol


}
\author{
Kirsten van Kessel \email{k.d.vankessel@students.uu.nl}
}
